+++ crdtree/src/CRDTree.ts
@@ -1,5 +1,5 @@
 import State from "./State";
-import {ICRDTree, ID, Index} from "./types/Types";
+import {ICRDTree, ID, Index} from "./API";
 import {ROOT} from "./Constants";
 import {BackendChange, Change} from "./Change";
 import {ActionKind, FrontendAction} from "./types/BaseAction";
