+++ crdtree/src/Constants.ts
@@ -1,3 +1,7 @@
+type ROOT = "ROOT";
+type ROOT_PARENT = "PARENT";
+type HEAD = "HEAD";
+
 const ROOT = "ROOT";
 const ROOT_PARENT = "PARENT";
 const HEAD = "HEAD";
