+++ crdtree/src/StateObject.ts
@@ -1,5 +1,5 @@
-import {ID, Index} from "./types/Types";
-import {BasePrimitive, ObjectKind} from "./types/Primitive";
+import {ID, Index} from "./API";
+import {BasePrimitive, ObjectKind} from "./Primitive";
 
 type Entry = { name: ID, kind: ObjectKind, value: BasePrimitive | ID, deleted: boolean };
 type MetaObject = Map<Index, Entry> | Array<Entry>;
